requirements.txt
SQLAlchemy
alembic 
python-dotenv
psycopg2-binary